requirements.txt
python-dotenv==1.0.0
openai==0.28.0
tiktoken==0.5.1
streamlit==1.28.2